+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Changed
 
+- Publish-results reminder emails are only sent to initiators involved in the project (admin log).
 - Update a4 for confidential poll questions
 - Poll dashboard and detail styling for confidential questions